[fix][dingo-exec] Calculate distance invoke by toolservice
