+++ dingo-calcite/src/main/java/io/dingodb/calcite/visitor/function/DingoGetVectorByDistanceVisitFun.java
@@ -80,7 +80,6 @@ public final class DingoGetVectorByDistanceVisitFun {
             VectorPointDistanceOperator operator = new VectorPointDistanceOperator(
                 distributions.firstEntry().getValue(),
                 rel.getVectorIndex(),
-                tableInfo.getId(),
                 rel.getIndexTableId(),
                 targetVector,
                 dimension,
